# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoCare API — vehicle owners book services, service centers fulfill them.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-centers:
    get:
      operationId: listServiceCenters
      tags: [serviceCenters]
      summary: Browse service centers
      parameters:
        - { name: specialty, in: query, required: false, schema: { type: string } }
        - { name: includeInactive, in: query, required: false, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ServiceCenter" }
  /service-centers/{centerId}:
    get:
      operationId: getServiceCenter
      tags: [serviceCenters]
      parameters:
        - { name: centerId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceCenter" }
    patch:
      operationId: updateServiceCenter
      tags: [serviceCenters]
      summary: Admin — suspend or reactivate a service center
      parameters:
        - { name: centerId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateActiveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceCenter" }
    delete:
      operationId: deleteServiceCenter
      tags: [serviceCenters]
      summary: Admin — permanently delete a service center (only if it has no bookings)
      parameters:
        - { name: centerId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "409":
          description: Conflict — center has dependent records, suspend instead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteConflict" }
...
  /service-centers/{centerId}/mechanics:
    get:
      operationId: listMechanicsForCenter
      tags: [mechanics]
      parameters:
        - { name: centerId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Mechanic" }
  /mechanics/{mechanicId}:
    get:
      operationId: getMechanic
      tags: [mechanics]
      parameters:
        - { name: mechanicId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mechanic" }
    patch:
      operationId: updateMechanic
      tags: [mechanics]
      summary: Admin — suspend or reactivate a mechanic
      parameters:
        - { name: mechanicId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateActiveInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mechanic" }
    delete:
      operationId: deleteMechanic
      tags: [mechanics]
      summary: Admin — permanently delete a mechanic (only if not assigned to any booking)
      parameters:
        - { name: mechanicId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "409":
          description: Conflict — mechanic has dependent records, suspend instead
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteConflict" }
...
  /service-centers/{centerId}/settings:
    patch:
      operationId: updateServiceCenterSettings
      tags: [serviceCenters]
      summary: Update center self-service settings (WhatsApp notifications, etc.)
      parameters:
        - { name: centerId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateServiceCenterSettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceCenter" }
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      parameters:
        - { name: ownerPhone, in: query, required: true, schema: { type: string } }
        - { name: unreadOnly, in: query, required: false, schema: { type: boolean } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - { name: notificationId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
  /notifications/mark-all-read:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkAllNotificationsReadInput" }
      responses:
        "204": { description: Marked }
  /notifications/generate-reminders:
    post:
      operationId: generateReminderNotifications
      tags: [notifications]
      summary: Run the service-reminder generator on demand
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateReminderNotificationsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [created]
                properties:
                  created: { type: integer }
  /push/vapid-public-key:
    get:
      operationId: getVapidPublicKey
      tags: [push]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [publicKey]
                properties:
                  publicKey: { type: string }
        "503":
          description: Push not configured
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
  /push/subscriptions:
    post:
      operationId: createPushSubscription
      tags: [push]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePushSubscriptionInput" }
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PushSubscription" }
    delete:
      operationId: deletePushSubscription
      tags: [push]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeletePushSubscriptionInput" }
      responses:
        "204": { description: Unsubscribed }
components:
  schemas:
...
    ServiceCenter:
      type: object
      required: [id, name, address, phone, specialties, rating, reviewsCount, openJobs]
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        city: { type: string }
        region: { type: string }
        phone: { type: string }
        specialties:
          type: array
          items: { type: string }
        rating: { type: number }
        reviewsCount: { type: integer }
        openJobs: { type: integer }
        imageUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        active: { type: boolean }
        whatsappOptIn: { type: boolean }
    Mechanic:
      type: object
      required: [id, serviceCenterId, name, yearsExperience, specialization, rating, completedJobs]
      properties:
        id: { type: string }
        serviceCenterId: { type: string }
        name: { type: string }
        yearsExperience: { type: integer }
        specialization: { type: string }
        certifications:
          type: array
          items: { type: string }
        rating: { type: number }
        completedJobs: { type: integer }
        avatarUrl: { type: string, nullable: true }
        active: { type: boolean }
...
    UpdateActiveInput:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
    DeleteConflict:
      type: object
      required: [error, reason]
      properties:
        error: { type: string }
        reason: { type: string, enum: [has_dependents] }
        details: { type: string }
...
    UpdateServiceCenterSettingsInput:
      type: object
      properties:
        whatsappOptIn: { type: boolean }
    Notification:
      type: object
      required: [id, ownerPhone, kind, title, body, dedupeKey, createdAt]
      properties:
        id: { type: string }
        ownerPhone: { type: string }
        kind: { type: string }
        title: { type: string }
        body: { type: string }
        vehicleId: { type: string, nullable: true }
        bookingId: { type: string, nullable: true }
        dedupeKey: { type: string }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    MarkAllNotificationsReadInput:
      type: object
      required: [ownerPhone]
      properties:
        ownerPhone: { type: string }
    GenerateReminderNotificationsInput:
      type: object
      properties:
        triggeredBy: { type: string }
    PushSubscription:
      type: object
      required: [id, ownerPhone, endpoint, p256dh, auth, createdAt]
      properties:
        id: { type: string }
        ownerPhone: { type: string }
        endpoint: { type: string }
        p256dh: { type: string }
        auth: { type: string }
        userAgent: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreatePushSubscriptionInput:
      type: object
      required: [ownerPhone, endpoint, p256dh, auth]
      properties:
        ownerPhone: { type: string, minLength: 1 }
        endpoint: { type: string, minLength: 1 }
        p256dh: { type: string, minLength: 1 }
        auth: { type: string, minLength: 1 }
        userAgent: { type: string, nullable: true }
    DeletePushSubscriptionInput:
      type: object
      required: [endpoint]
      properties:
        endpoint: { type: string, minLength: 1 }