# lib/api-zod/src/generated/types/createPushSubscriptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoCare API — vehicle owners book services, service centers fulfill them.
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePushSubscriptionInput {
  /** @minLength 1 */
  ownerPhone: string;
  /** @minLength 1 */
  endpoint: string;
  /** @minLength 1 */
  p256dh: string;
  /** @minLength 1 */
  auth: string;
  userAgent?: string | null;
}