# lib/api-zod/src/generated/types/deletePushSubscriptionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoCare API — vehicle owners book services, service centers fulfill them.
 * OpenAPI spec version: 0.1.0
 */

export interface DeletePushSubscriptionInput {
  /** @minLength 1 */
  endpoint: string;
}